# model/member_chunk: optimise deserialisation