--- a/model/benches/member_chunk.rs
+++ b/model/benches/member_chunk.rs
@@ -1,4 +1,4 @@
-use criterion::{Criterion, criterion_group, criterion_main};
+use criterion::{criterion_group, criterion_main, Criterion};
 
 use twilight_model::gateway::payload::MemberChunk;