--- a/model/src/guild/member.rs
+++ b/model/src/guild/member.rs
@@ -118,7 +118,6 @@ impl<'de> Visitor<'de> for MemberMapDeserializerVisitor {
         f.write_str("a sequence of members")
     }
 
-    #[allow(unused)]
     fn visit_seq<S: SeqAccess<'de>>(self, mut seq: S) -> Result<Self::Value, S::Error> {
         let mut map = seq
             .size_hint()