@BlackHoleFox 2020-06-24
Would you consider using `criterion` instead of the Nightly only test feature?